+++ Daedalus/Rotation/AsclepiusCore/Modules/ResurrectionModule.cs
@@ -23,7 +23,7 @@ public sealed class ResurrectionModule : BaseResurrectionModule<IAsclepiusContex
     protected override int RaiseMpCost => 2400;
 
     protected override IBattleChara? FindDeadPartyMemberNeedingRaise(IAsclepiusContext context)
-        => context.PartyHelper.FindDeadPartyMemberNeedingRaise(context.Player);
+        => context.PartyHelper.FindDeadPartyMemberNeedingRaise(context.Player, context.Configuration.Resurrection.RaiseAllianceMembers);
 
     protected override bool HasSwiftcast(IAsclepiusContext context) => context.HasSwiftcast;